requirements.txt
Django>=3.2.0
djangorestframework